-Get tailwind components working with a fresh template
