+++ examples/encrypted-chat/index.html
@@ -5,6 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + React + TS</title>
+    <link rel="stylesheet" href="https://rsms.me/inter/inter.css">
   </head>
   <body>
     <div id="root"></div>
