+++ examples/encrypted-chat/src/index.css
@@ -1,4 +1,8 @@
-:root {
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+/* :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
@@ -79,4 +83,4 @@ input {
 }
 input:focus {
   border-color: #747bff;
-}
+} */
