+++ examples/encrypted-chat/tailwind.config.js
@@ -1,0 +1,6 @@
+export default {
+  content: [
+    "./src/**/*.{html,js,tsx}",
+  ],
+  plugins: [],
+};
